Basic handling of pushes.
The general tactic is to store the revision at which we last performed
a sync and, when performing a new sync, divide the added commits into
PRs that landed upstream, and for each such PR copy the upstream at
the time that landed, before reapplying any unmerged local changes,
and then cherry-picking metadata that we created for the PR (if
any). This copying approach means that we can't diverge too far from
upstream by accident.