Add working docker-compose/celery support.
We filter messages we are interested in and put them in an internal
rabbitmq queue. We have a celery worker that processes these messages
one at a time, as they become avaiable.

Docker-wise we arrange things so that there is one container per
process:

sync - Listens to Pulse messages and enques the interesting ones.
rabbitmq - Runs the rabbitmq server
celery - Runs the celery process that processes the actual messages

Presently we mount the repo directory into the containers so that we
are able to work on the code without requiring rebuilds. There are
some permissions that need to be manually configured at present, and
the setup doesn't automatically generate the models or clone
repositories.