Move to using one Landing object per time we land  from upstream.
Instead of having a single Landing object updated to record the commit
of the previous landing, have one per time we attempt to land, and
link to WptCommit objects to record which commits were landed in each
landing.

This also adds an initial status field to the Landing table which will
be used with the next phase, which is to split these tasks into phases
that can be resumed after a failure.