Rewite landing code to work in terms of discrete steps.
So the idea here is that when we get an event we run a pipeline that
is more or less composed of a set of steps (but that composition isn't
declarative, so we should be able to do things like loop over
steps). Then the goal is that anything that was in progress that
failed should be able to run through the complete pipeline, loading
data or skiping steps according to its state, and eventually get to
the point where it failed and continue. That isn't really enforced;
you need to ensure that you don't accidentially change or duplicate
state when rerunning a step. But what they give you is convenient sync
points at which the database state is saved, and a declarative way to
manage state transitions.