Store the commits that we have applied to a landing.
This means that when we have e.g. a merge conflict, we can resume and
work on the next commit in the sequence without having to complete the
entire set of landings and move on to the next step.

In an ideal world we could instead work out from the state of vcs
which commits are already applied, but that's harder, so it makes some
sense to start with this model.