Refactor statuses and apply to upstreaming.
Instead of having a single status per object covering its whole
lifecycle, we have a simplified overall status of
active/completed/aborted plus various attribures that communicate bits
of state. This has the advantage that we don't need to model
things (particularly retries) as linear progressions through a fixed
set of states, but can indicate exactly which steps need to be rerun
in order to retry the landing process.