Add support for removing stale entries in the moz.build file.
If there are stale entries in the moz.build file we will fail make
check tests. This patch uses lib2to3 to parse that file and match
statements of the form

with File(<arg>):

It gathers all the values of <arg> and checks that they match at least
one filesystem path. Any that don't match are removed.

Ideally this would be improved to take account of test moves, but that
is a little involved since we need to figure out which entries
correspond to a moved directory, and what the appropriate pattern is
to match the destination.